pyproject.toml
[project]
name = "binance-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">3.12, <3.13"
dependencies = [
    "gospeed-api>=2.0.6",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "joblib>=1.4.2",
    "loguru>=0.7.3",
    "numpy>=2.2.2",
    "polars>=1.20.0",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
    "xmltodict>=0.14.2",
]
